Extract route paths to break import cycle
